fix: 카카오로그인 추가정보받기

+++ src/common/api/baseApi.js
@@ -4,7 +4,7 @@ const baseApi = axios.create({
 	headers: {
 		'Content-Type': 'application/json',
 	},
-	baseURL: process.env.NODE_ENV === 'production' ? "" : 'http://localhost:33001',
+	baseURL: process.env.NODE_ENV === 'production' ? "" : 'http://localhost:33000',
 	withCredentials: true,
 	timeout: 1000 * 30,
 });
@@ -17,6 +17,8 @@ baseApi.interceptors.request.use((config) => {
 	return config;
 });
 
+let isRedirectiong = false;
+
 baseApi.interceptors.response.use();
 
 baseApi.interceptors.response.use(
@@ -26,11 +28,12 @@ baseApi.interceptors.response.use(
 		return response;
 	},
 	function (error) {
-		if (error.status === 401) {
+		if (error.status === 401 && !isRedirectiong) {
+			isRedirectiong = true;
 			alert('올바른 로그인 정보가 아닙니다. 로그인 페이지로 이동합니다.');
 			localStorage.clear();
 			window.location.replace('/');
-		}	
+		}
 		// Any status codes that falls outside the range of 2xx cause this function to trigger
 		// Do something with response error
 		return Promise.reject(error);
